Basic gps functionality.
It uses both fine and course grain location which also requires internet.  The permissions are declared in the manifest.  We should incorporate this into A0 since it can prompt the user each time they log on for GPS to be turned on.  To do that we just need to add to the provider disabled function.